requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
openai
gtts
SpeechRecognition
PyPDF2
python-docx
pandas
langchain_community
